+++ Cargo.toml
@@ -22,7 +22,7 @@ targets = [
 ]
 
 [features]
-default = [ "file-drop", "objc-exception", "protocol", "tao", "rwh_05" ]
+default = [ "file-drop", "objc-exception", "protocol" ]
 objc-exception = [ "objc/exception" ]
 file-drop = [ ]
 protocol = [ ]
@@ -40,32 +40,20 @@ serde = { version = "1.0", features = [ "derive" ] }
 serde_json = "1.0"
 thiserror = "1.0"
 url = "2.4"
-tao = { version = "0.23", default-features = false, features = [ "serde" ], optional = true }
-http = "0.2.9"
-rwh_04 = { package = "raw-window-handle", version = "0.4", optional = true }
-rwh_05 = { package = "raw-window-handle", version = "0.5", features = ["std"], optional = true }
-rwh_06 = { package = "raw-window-handle", version = "0.6", features = ["std"], optional = true }
-
-[dev-dependencies]
-http-range = "0.1.5"
-base64 = "0.21"
+http = "0.2"
+raw-window-handle= { version = "0.6", features = ["std"] }
 
 [target."cfg(any(target_os = \"linux\", target_os = \"dragonfly\", target_os = \"freebsd\", target_os = \"openbsd\", target_os = \"netbsd\"))".dependencies]
 javascriptcore-rs = { version = "=1.1.1", features = [ "v2_28" ] }
 webkit2gtk = { version = "=2.0", features = [ "v2_38" ] }
 webkit2gtk-sys = "=2.0"
-gio = "0.18"
-glib = "0.18"
 gtk = "0.18"
-gdk = "0.18"
 soup3 = "0.5"
-winit = { version = "0.29", package = "winit-gtk", features = [ "serde" ], optional = true }
 
 [target."cfg(target_os = \"windows\")".dependencies]
 webview2-com = "0.27"
 windows-implement = "0.51"
 dunce = "1"
-winit = { version = "0.28", features = [ "serde" ], optional = true }
 
   [target."cfg(target_os = \"windows\")".dependencies.windows]
   version = "0.51"
@@ -93,7 +81,6 @@ cocoa = "0.25"
 core-graphics = "0.23"
 objc = "0.2"
 objc_id = "0.1"
-winit = { version = "0.28", features = [ "serde" ], optional = true }
 
 [target."cfg(target_os = \"android\")".dependencies]
 crossbeam-channel = "0.5"
@@ -101,3 +88,8 @@ html5ever = "0.26"
 kuchiki = { package = "kuchikiki", version = "0.8" }
 sha2 = "0.10"
 base64 = "0.21"
+
+[dev-dependencies]
+http-range = "0.1.5"
+tao = { git = "https://github.com/tauri-apps/tao", branch = "rwh" }
+winit = "0.29"
