+++ examples/custom_protocol.rs
@@ -8,15 +8,14 @@ use std::{
 };
 
 use http::Request;
-use rwh_05::HasRawWindowHandle;
+use tao::{
+  event::{Event, WindowEvent},
+  event_loop::{ControlFlow, EventLoop},
+  window::WindowBuilder,
+};
 use wry::{
-  application::{
-    event::{Event, StartCause, WindowEvent},
-    event_loop::{ControlFlow, EventLoop},
-    window::WindowBuilder,
-  },
   http::{header::CONTENT_TYPE, Response},
-  webview::WebViewBuilder,
+  WebViewBuilder,
 };
 
 const PAGE1_HTML: &[u8] = include_bytes!("custom_protocol_page1.html");
@@ -28,8 +27,7 @@ fn main() -> wry::Result<()> {
     .build(&event_loop)
     .unwrap();
 
-  let _webview = WebViewBuilder::new(window.raw_window_handle())
-    .unwrap()
+  let _webview = WebViewBuilder::new(&window)
     .with_asynchronous_custom_protocol("wry".into(), move |request, responder| {
       match get_wry_response(request) {
         Ok(http_response) => responder.respond(http_response),
@@ -49,13 +47,12 @@ fn main() -> wry::Result<()> {
   event_loop.run(move |event, _, control_flow| {
     *control_flow = ControlFlow::Wait;
 
-    match event {
-      Event::NewEvents(StartCause::Init) => println!("Wry application started!"),
-      Event::WindowEvent {
-        event: WindowEvent::CloseRequested,
-        ..
-      } => *control_flow = ControlFlow::Exit,
-      _ => (),
+    if let Event::WindowEvent {
+      event: WindowEvent::CloseRequested,
+      ..
+    } = event
+    {
+      *control_flow = ControlFlow::Exit
     }
   });
 }
