+++ examples/dragndrop.rs
@@ -2,18 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use rwh_05::HasRawWindowHandle;
+use tao::{
+  event::{Event, WindowEvent},
+  event_loop::{ControlFlow, EventLoop},
+  window::WindowBuilder,
+};
+use wry::WebViewBuilder;
 
 fn main() -> wry::Result<()> {
-  use wry::{
-    application::{
-      event::{Event, StartCause, WindowEvent},
-      event_loop::{ControlFlow, EventLoop},
-      window::WindowBuilder,
-    },
-    webview::WebViewBuilder,
-  };
-
   const HTML: &str = r#"data:text/html,
 Drop files onto the window and read the console!<br>
 Dropping files onto the following form is also possible:<br><br>
@@ -22,8 +18,7 @@ Dropping files onto the following form is also possible:<br><br>
 
   let event_loop = EventLoop::new();
   let window = WindowBuilder::new().build(&event_loop).unwrap();
-  let _webview = WebViewBuilder::new(window.raw_window_handle())
-    .unwrap()
+  let _webview = WebViewBuilder::new(&window)
     .with_url(HTML)?
     .with_file_drop_handler(|data| {
       println!("Window 1: {:?}", data);
@@ -34,13 +29,12 @@ Dropping files onto the following form is also possible:<br><br>
   event_loop.run(move |event, _, control_flow| {
     *control_flow = ControlFlow::Wait;
 
-    match event {
-      Event::NewEvents(StartCause::Init) => println!("Wry application started!"),
-      Event::WindowEvent {
-        event: WindowEvent::CloseRequested,
-        ..
-      } => *control_flow = ControlFlow::Exit,
-      _ => {}
+    if let Event::WindowEvent {
+      event: WindowEvent::CloseRequested,
+      ..
+    } = event
+    {
+      *control_flow = ControlFlow::Exit
     }
   });
 }
