+++ examples/form.rs
@@ -7,15 +7,15 @@ use std::{
   fs::{canonicalize, read},
 };
 
-use rwh_05::HasRawWindowHandle;
+use tao::{
+  event::{Event, WindowEvent},
+  event_loop::{ControlFlow, EventLoop},
+  window::WindowBuilder,
+};
+
 use wry::{
-  application::{
-    event::{Event, StartCause, WindowEvent},
-    event_loop::{ControlFlow, EventLoop},
-    window::WindowBuilder,
-  },
   http::{header::CONTENT_TYPE, method::Method, Response},
-  webview::WebViewBuilder,
+  WebViewBuilder,
 };
 
 fn main() -> wry::Result<()> {
@@ -25,8 +25,7 @@ fn main() -> wry::Result<()> {
     .build(&event_loop)
     .unwrap();
 
-  let _webview = WebViewBuilder::new(window.raw_window_handle())
-    .unwrap()
+  let _webview = WebViewBuilder::new(&window)
     .with_custom_protocol("wry".into(), move |request| {
       if request.method() == Method::POST {
         let body_string = String::from_utf8_lossy(request.body());
@@ -53,13 +52,12 @@ fn main() -> wry::Result<()> {
   event_loop.run(move |event, _, control_flow| {
     *control_flow = ControlFlow::Wait;
 
-    match event {
-      Event::NewEvents(StartCause::Init) => println!("Wry application started!"),
-      Event::WindowEvent {
-        event: WindowEvent::CloseRequested,
-        ..
-      } => *control_flow = ControlFlow::Exit,
-      _ => (),
+    if let Event::WindowEvent {
+      event: WindowEvent::CloseRequested,
+      ..
+    } = event
+    {
+      *control_flow = ControlFlow::Exit
     }
   });
 }
