+++ examples/new_window_req_event.rs
@@ -2,22 +2,18 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use rwh_05::HasRawWindowHandle;
+use tao::{
+  event::{Event, WindowEvent},
+  event_loop::{ControlFlow, EventLoopBuilder},
+  window::WindowBuilder,
+};
+use wry::WebViewBuilder;
+
+enum UserEvent {
+  NewWindow(String),
+}
 
 fn main() -> wry::Result<()> {
-  use wry::{
-    application::{
-      event::{Event, StartCause, WindowEvent},
-      event_loop::{ControlFlow, EventLoopBuilder},
-      window::WindowBuilder,
-    },
-    webview::WebViewBuilder,
-  };
-
-  enum UserEvent {
-    NewWindow(String),
-  }
-
   let html = r#"
     <body>
       <div>
@@ -32,8 +28,9 @@ fn main() -> wry::Result<()> {
   let proxy = event_loop.create_proxy();
   let window = WindowBuilder::new()
     .with_title("Hello World")
-    .build(&event_loop)?;
-  let _webview = WebViewBuilder::new(window.raw_window_handle())?
+    .build(&event_loop)
+    .unwrap();
+  let _webview = WebViewBuilder::new(&window)
     .with_html(html)?
     .with_new_window_req_handler(move |uri: String| {
       let submitted = proxy.send_event(UserEvent::NewWindow(uri.clone())).is_ok();
@@ -46,7 +43,6 @@ fn main() -> wry::Result<()> {
     *control_flow = ControlFlow::Wait;
 
     match event {
-      Event::NewEvents(StartCause::Init) => println!("Wry has started!"),
       Event::WindowEvent {
         event: WindowEvent::CloseRequested,
         ..
