+++ examples/user_agent.rs
@@ -2,18 +2,14 @@
 // SPDX-License-Identifier: Apache-2.0
 // SPDX-License-Identifier: MIT
 
-use rwh_05::HasRawWindowHandle;
+use tao::{
+  event::{Event, WindowEvent},
+  event_loop::{ControlFlow, EventLoop},
+  window::WindowBuilder,
+};
+use wry::{webview_version, WebViewBuilder};
 
 fn main() -> wry::Result<()> {
-  use wry::{
-    application::{
-      event::{Event, StartCause, WindowEvent},
-      event_loop::{ControlFlow, EventLoop},
-      window::WindowBuilder,
-    },
-    webview::{webview_version, WebViewBuilder},
-  };
-
   let current_version = env!("CARGO_PKG_VERSION");
   let current_webview_version = webview_version().unwrap();
   let user_agent_string = format!(
@@ -26,8 +22,9 @@ fn main() -> wry::Result<()> {
   let event_loop = EventLoop::new();
   let window = WindowBuilder::new()
     .with_title("Hello World")
-    .build(&event_loop)?;
-  let _webview = WebViewBuilder::new(window.raw_window_handle())?
+    .build(&event_loop)
+    .unwrap();
+  let _webview = WebViewBuilder::new(&window)
     .with_user_agent(&user_agent_string)
     .with_url("https://www.whatismybrowser.com/detect/what-is-my-user-agent")?
     .build()?;
@@ -35,13 +32,12 @@ fn main() -> wry::Result<()> {
   event_loop.run(move |event, _, control_flow| {
     *control_flow = ControlFlow::Wait;
 
-    match event {
-      Event::NewEvents(StartCause::Init) => println!("Wry has started!"),
-      Event::WindowEvent {
-        event: WindowEvent::CloseRequested,
-        ..
-      } => *control_flow = ControlFlow::Exit,
-      _ => (),
+    if let Event::WindowEvent {
+      event: WindowEvent::CloseRequested,
+      ..
+    } = event
+    {
+      *control_flow = ControlFlow::Exit
     }
   });
 }
