+++ src/android/kotlin/Logger.kt
