+++ src/android/kotlin/WryActivity.kt
